exam_m7/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.4
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
generic==1.1.2
generics==7.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.5.0
typing==3.7.4.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
